Make OVC manage the IP Addresses of a cloudspace
Do not rely on routeros to assign IPs based on a pool as it has proven
time over time again that its unreliable at keeping track of its leases.

Instead we assign the IPs and macaddresses inside OVC and configured
those leaees `manually` inside routeros.

This requires a migration script to drop the pool in the
dhcpserver for existing routeroses and loop al vms with IP `undefined`
to be assigned by the new getFreeIPAddress method.

Signed-off-by: Jo De Boeck <deboeck.jo@gmail.com>